% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cdf_from_obs}
\alias{cdf_from_obs}
\title{cdf_from_obs}
\usage{
cdf_from_obs(obs)
}
\arguments{
\item{obs}{A numeric vector of observations}
}
\value{
An XPtr to an EmpiricalCDF object
}
\description{
Create an empirical CDF from observations
}
