% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cdf_from_timing}
\alias{cdf_from_timing}
\title{cdf_from_timing}
\usage{
cdf_from_timing(t, obs)
}
\arguments{
\item{t}{A numeric vector of timings}

\item{obs}{A numeric vector of dosages}
}
\description{
Create an empirical CDF from timings and dosages
}
