% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eval_cdf}
\alias{eval_cdf}
\title{eval_cdf}
\usage{
eval_cdf(cdf, x)
}
\arguments{
\item{cdf}{An XPtr to an EmpiricalCDF object}

\item{x}{The point at which to evaluate the CDF}
}
\description{
Evaluate the CDF at a given point
}
