% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wasserstein}
\alias{wasserstein}
\title{wasserstein}
\usage{
wasserstein(cdf1, cdf2)
}
\arguments{
\item{cdf1}{An XPtr to an EmpiricalCDF object}

\item{cdf2}{An XPtr to an EmpiricalCDF object}
}
\value{
The 1-Wasserstein distance between the two CDFs
}
\description{
Computes the 1-Wasserstein metric between two empirical CDFs
}
